Cleared diffusers library
