+++ image_generation/common/imwrite/include/imwrite.hpp
