+++ image_generation/common/imwrite/src/imwrite.cpp
